my-telegram-bot/requirements.txt
aiogram==3.21.0
aiohttp==3.9.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
feedparser==6.0.10
